+++ Leave_Application/Leave_Application/Program.cs
@@ -1,21 +1,21 @@
 using Leave_Application.Data;
 using Microsoft.EntityFrameworkCore;
-using Leave_Application.Data; 
+using Leave_Application.Services;
 
 var builder = WebApplication.CreateBuilder(args);
 
-// Add services to the container.
+
 builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
-// Add DbContext using the connection string from appsettings.json
+
 builder.Services.AddDbContext<AppDbContext>(options =>
     options.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection")));
 
 var app = builder.Build();
 
-// Configure the HTTP request pipeline.
+
 if (app.Environment.IsDevelopment())
 {
     app.UseSwagger();
