+++ Leave_Application/Leave_Application/bin/Debug/net8.0/Leave_Application.staticwebassets.endpoints.json
@@ -1,5 +1,1 @@
-{
-  "Version": 1,
-  "ManifestType": "Build",
-  "Endpoints": []
-}
+{"Version":1,"ManifestType":"Build","Endpoints":[]}
